requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
pyocr>=0.8.0
requests>=2.25.0
Flask>=2.0.0
Werkzeug>=2.0.0
